Modify error handling in HTTP gate
sendResponse primitive must be called only once during the processing
of an HTTP request. Previous code didn't follow this rule and browsers
weren't able to decode the response due to wrong Content-lenght header.

This commit introduces SPARQL11ProtocolException which is trown inside
the parse method so the sendResponse is called only once in catch clause
of the base class SPARQL11Handler