Fix webtoken expiration check
With this change SEPA now checks correctly the expiration date of a
JSONWebToken. Notice that the expiration date is stored as seconds
from Unix time, not milliseconds. This error was causing deadlocks
in the token regeneration flow.
i.e.
- token expiration time 12:00:00.800
- real expiration time stored 12:00:00
using the webToken at 12:00:00.000 results in a "token is expired" error.
trying to regenerate it also cause an error: token is still valid and it
will expire at 12:00:00.800

This change covertes all the token dates to seconds from Unix time